eyes-sdk-core, eyes-selenium: moved branch functions to EyesRunner and removed batchId from getting SCM info
