+++ packages/eyes-selenium/lib/runner/EyesRunner.js
@@ -1,5 +1,7 @@
 'use strict'
 
+const {makeGetBatchInfo, getScmInfo} = require('@applitools/eyes-sdk-core')
+
 class EyesRunner {
   constructor() {
     /** @type {Eyes[]} */
@@ -33,6 +35,26 @@ class EyesRunner {
 
       await Promise.all(promises)
     }
+  }
+
+  makeGetBatchInfo(fetchBatchInfo) {
+    if (!this._getBatchInfo) {
+      this._getBatchInfo = makeGetBatchInfo(fetchBatchInfo)
+    }
+  }
+
+  async getBatchInfoWithCache(batchId) {
+    if (this._getBatchInfo) {
+      return this._getBatchInfo(batchId)
+    } else {
+      throw new Error(
+        'Eyes runner could not get batch info since makeGetBatchInfo was not called before',
+      )
+    }
+  }
+
+  async getScmInfoWithCache(...args) {
+    return getScmInfo(...args)
   }
 }
